Return a RecognitionResult when all images sent are invalid.
This is preferable to throwing an exception because it is consistent with
the partial error case where only some of the images are invalid.
Clients are expected to iterate through the results of an batch recognition
request, checking the status code of each, and this allows them to do that.